Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/workers/Worker_ErrorEvent_type-expected.txt

CONSOLE MESSAGE: Error: Error Message

PASS  AbstractWorker ErrorEvent.type
